Upgrade the example operator to modal
I put a lot of work into this to make it advanced and simple at the same time.
I'm quite happy with how I've structured the code.

Features:
 - Dynamic tooltip
 - Custom header text with values
 - Custom status bar with key icons
 - Usable as modal and with undo redo
 - Reliable cursor warping

One thing that's missing is error catching and cleanup of the status bar and header.
I'll look into it later, calling it a night now.