Manually set and clear status and header
Status only needs to be set in invoke, because it sets it to a function.
Header needs to be set at the end of each modal run, because it's just static text.
I got rid of the utility function for clearing status and header, so I do that manually at the top of cleanup now, because it's more transparent.