Make alt pass through in modal method
You'll still be able to use alt + key in the modal because it checks for the event's ascii.
I'm adding this because some people use alt left click to navigate (and I don't mean middle mouse emulation in blender, but manually changing the keymap for it).
And it's nice being able to reposition the cursor with this I suppose.