Simplify utils.safety.respond
Instead of ignoring it when self.cancel doesn't exist, it adds that to the exception.
I think this makes more sense, because without a cancel method, there's no point in using the decorator in the first place.
I'm using try except else, which means try the code, do something with the exception, or if there is no exception, do the else block.
It looks like I could pass the exception instead and just have one line with self.report at the end, but that doesn't work, because it won't include the exception from the failed cancel call.