Create utils.draw_2d module
This is a function that draws two columns of text.
The left one is right aligned, the right one is left aligned.
You can align the entire block to any side, corner, or center.
I might make distance between lines and such parameters in the future.
And perhaps an easy way to add a box around the text.