Remove safety decorator from execute
I've come to the conclusion that this is actually a bad thing.
Because execute was wrapped, exceptions wouldn't occur inside modal, so the operator would perform cleanup, but not exit.
The whole point of the decorator is to remove draw handlers and restore the scene to how it was, which is not relevant to execute.
When execute is being called from the redo panel, there shouldn't be draw handlers around anymore, so those don't need cleanup.
And with the redo panel it uses blender's undo, so there's no point in restoring object states manually.
I suppose one solution would have been to check for the return from execute in modal, but just letting the wrapper I put on modal handle it is easier.
In short, don't use the safety decorator on execute.