Add type hinting to utility functions
Type hinting is pretty useful.
The only one that bothers me is for draw_text I made horiztonal and vertical just str instead of specifying which str values are allowed. I don't know a proper way of doing this unfortunately.
I don't like using an Enum because I want to be able to use strings for these parameters.